Disable lastinsertid in go1.10 and later (#510)
* Update issue templates

* Delete custom.md

* Delete issue-template.md

* Update issue templates

* disable lastinsertid in go1.10 and later
