+++ queries_test.go
@@ -777,56 +777,6 @@ func TestAffectedRows(t *testing.T) {
 	}
 }
 
-func TestIdentity(t *testing.T) {
-	conn := open(t)
-	defer conn.Close()
-
-	tx, err := conn.Begin()
-	if err != nil {
-		t.Fatal("Begin tran failed", err)
-	}
-	defer tx.Rollback()
-
-	res, err := tx.Exec("create table #foo (bar int identity, baz int unique)")
-	if err != nil {
-		t.Fatal("create table failed")
-	}
-
-	res, err = tx.Exec("insert into #foo (baz) values (1)")
-	if err != nil {
-		t.Fatal("insert failed")
-	}
-	n, err := res.LastInsertId()
-	if err != nil {
-		t.Fatal("last insert id failed")
-	}
-	if n != 1 {
-		t.Error("Expected 1 for identity, got ", n)
-	}
-
-	res, err = tx.Exec("insert into #foo (baz) values (20)")
-	if err != nil {
-		t.Fatal("insert failed")
-	}
-	n, err = res.LastInsertId()
-	if err != nil {
-		t.Fatal("last insert id failed")
-	}
-	if n != 2 {
-		t.Error("Expected 2 for identity, got ", n)
-	}
-
-	res, err = tx.Exec("insert into #foo (baz) values (1)")
-	if err == nil {
-		t.Fatal("insert should fail")
-	}
-
-	res, err = tx.Exec("insert into #foo (baz) values (?)", 1)
-	if err == nil {
-		t.Fatal("insert should fail")
-	}
-}
-
 func queryParamRoundTrip(db *sql.DB, param interface{}, dest interface{}) {
 	err := db.QueryRow("select ?", param).Scan(dest)
 	if err != nil {
